requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
defusedxml==0.7.1
dj-rest-auth==4.0.0
Django==4.2
django-allauth==0.54.0
django-ckeditor==6.5.1
django-dotenv==1.4.2
django-js-asset==2.0.0
django-rest-auth==0.9.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
greenlet==2.0.2
idna==3.4
msgpack==1.0.5
neovim==0.3.1
oauthlib==3.2.2
Pillow==9.5.0
pycparser==2.21
PyJWT==2.6.0
pynvim==0.4.3
python3-openid==3.2.0
pytz==2023.3
requests==2.30.0
requests-oauthlib==1.3.1
six==1.16.0
source==1.2.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.2
